rename file to usecase
