+++ numba/cuda/tests/cudapy/extensions_usecases.py
