+++ numba/cuda/tests/cudapy/test_localmem.py
@@ -3,7 +3,7 @@ import numpy as np
 from numba import cuda, int32, complex128, void
 from numba.core.errors import TypingError
 from numba.cuda.testing import unittest, CUDATestCase, skip_on_cudasim
-from .utils import test_struct_model_type, TestStruct
+from .extensions_usecases import test_struct_model_type, TestStruct
 
 
 def culocal(A, B):
