+++ numba/cuda/tests/cudapy/test_sm_creation.py
@@ -3,7 +3,7 @@ from numba import cuda, float32, int32, void
 from numba.core.errors import TypingError
 from numba.cuda.testing import unittest, CUDATestCase
 from numba.cuda.testing import skip_on_cudasim
-from .utils import test_struct_model_type
+from .extensions_usecases import test_struct_model_type
 
 GLOBAL_CONSTANT = 5
 GLOBAL_CONSTANT_2 = 6
